requirements.txt
streamlit
pandas
openpyxl
requests
pillow
pytesseract
transformers
torch
